Update CalendarHeatmap.js
Update to the CalendarHeatmap.js to add Day numubers to the squares.   New prop introduced usage as below;

dayNumberStyle={{
    fontSize: "8", // Customize font size
    color: "#333",  // Customize text color
}}
		  
In addition to the above, another prop was added to control when the day numbers would be shown.  If true, they will only show when there is a date in the value array.
  
showDayNumbersForEntriesOnly={true} // Set to true to only show day numbers where there is an entry